+++ src/components/menu.js
@@ -1,4 +1,4 @@
-import {elementTemplate} from "./utils";
+import {AbstractComponent} from "./abstract-component";
 
 /**
  * Функция, возвращающая разметку элемента меню
@@ -22,7 +22,7 @@ const menuState = {
   }
 };
 
-export class Menu extends elementTemplate {
+export class Menu extends AbstractComponent {
   constructor(events) {
     super();
     this._events = events;
